+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.2.4
+## Eliminating Query Builder
+- Removed the query builder functionality
+
 ## 1.2.3
 ### 🔥 Enhancements
 - Added `fields` parameter support to `fetchRecordsWithFilter`, allowing users to retrieve only specific fields along with filters.